extract tool call conversation into a new module
