+++ crates/goose/src/acp/server/load_session.rs
@@ -1,3 +1,7 @@
+use super::tool_calls::{
+    extract_tool_call_update_meta, pending_tool_call_from_request,
+    tool_call_update_fields_from_response,
+};
 use super::*;
 
 fn replay_audience_annotations(audience: &[Role]) -> Annotations {
