+++ crates/goose/src/acp/server/tool_calls/mod.rs
@@ -1,0 +1,6 @@
+mod conversion;
+
+pub(super) use conversion::{
+    extract_tool_call_update_meta, format_tool_name, pending_tool_call_from_request,
+    tool_call_identity_meta, tool_call_update_fields_from_response,
+};
